STUDENTMARKET/requirements.txt
bcrypt==5.0.0
bleach==6.1.0
blinker==1.9.0
click==8.3.1
Deprecated==1.3.1
dnspython==2.8.0
email-validator==2.1.0
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-Limiter==3.5.0
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-PyMongo==2.3.0
Flask-WTF==1.2.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.6.0
Markdown==3.5.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ordered-set==4.1.0
packaging==25.0
Pygments==2.19.2
pymongo==4.6.0
python-dateutil==2.8.2
python-dotenv==1.0.0
rich==13.9.4
six==1.17.0
typing_extensions==4.15.0
webencodings==0.5.1
Werkzeug==3.0.1
wrapt==2.0.1
WTForms==3.1.1
